+++ src/test/basic.test.js
@@ -83,5 +83,44 @@ describe('basic', () => {
       // $FlowFixMe
       expect(wrapper.node.ref).toNotExist()
     })
+
+    class InnerComponent extends Component {
+      render() {
+        return null
+      }
+    }
+
+    it('should not leak the innerRef prop to the wrapped child', () => {
+      const OuterComponent = styled(InnerComponent)``
+
+      class Wrapper extends Component {
+        testRef: any;
+
+        render() {
+          return <OuterComponent innerRef={(comp) => { this.testRef = comp }} />
+        }
+      }
+
+      const wrapper = mount(<Wrapper />)
+      const innerComponent = wrapper.find(InnerComponent).first()
+
+      // $FlowFixMe
+      expect(wrapper.node.testRef).toBe(innerComponent.node)
+      expect(innerComponent.prop('innerRef')).toNotExist()
+    })
+
+    it('should pass the full className to the wrapped child', () => {
+      const OuterComponent = styled(InnerComponent)``
+
+      class Wrapper extends Component {
+        render() {
+          return <OuterComponent className="test"/>
+        }
+      }
+
+      const wrapper = mount(<Wrapper />)
+      expect(wrapper.find(InnerComponent).prop('className'))
+        .toBe('test sc-c d')
+    })
   })
 })